+++ src/agent/web.ts
@@ -153,6 +153,7 @@ export default class Web extends WebAPI {
       ignore_requester_presence,
     });
   }
+
   /**
    * Removes a user from chat. Removing customer user type is not allowed. It's always possible to remove the requester from the chat.
    * @param chat_id - chat to remove user from
@@ -171,6 +172,7 @@ export default class Web extends WebAPI {
       ignore_requester_presence,
     });
   }
+
   /**
    * Sends an Event object. Use this method to send a message by specifing the Message event type in the request.
    * It's possible to write to a chat without joining it. The user sending an event will be automatically added to the chat
