+++ src/customer/structures/index.ts
@@ -1,0 +1,4 @@
+export * from "./events";
+export * from "./pushes";
+export * from "./responses";
+export * from "./structures";
